refactor: Rename `lib/resolveOutput.js`
