+++ lib/output-command.js
@@ -6,7 +6,7 @@ const log = require('./log');
 const isAuthenticated = require('./is-authenticated');
 const throwAuthError = require('./throwAuthError');
 const resolveOutputs = require('./resolveOutputs');
-const resolveOutput = require('./resolveOutput');
+const resolveOutput = require('./resolve-output');
 
 const resolveInput = function (ctx) {
   const {
