+++ lib/resolve-output.js
