+++ lib/variables.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const ensureString = require('type/string/ensure');
-const resolveOutput = require('./resolveOutput');
+const resolveOutput = require('./resolve-output');
 const resolveParams = require('./resolveParams');
 const throwAuthError = require('./throwAuthError');
